# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Projet API",
    "description": "Proof-of-work hiring infrastructure.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/public/x/{company_slug}": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Company Listing",
        "description": "FR-104 \u2014 one company's own programmes, e.g. for a careers page.\n\nA draft stays unreachable here too (FR-056) \u2014 the same rule as the\nsingle-programme page, just applied across the whole company.",
        "operationId": "company_listing_public_x__company_slug__get",
        "parameters": [
          {
            "name": "company_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Company Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicListingSummary"
                  },
                  "title": "Response Company Listing Public X  Company Slug  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/public/challenges": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Platform Directory",
        "description": "FR-105 \u2014 every active programme across companies, for a browse page.\n\nActive means the week is not over yet: both accepting applications (`open`)\nand still running with applications shut (`closed`). Complete programmes\ndrop off. The single-company listing above also shows complete ones, since\na company's own careers page is a different audience.",
        "operationId": "platform_directory_public_challenges_get",
        "parameters": [
          {
            "name": "role_slug",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role Slug"
            }
          },
          {
            "name": "cluster",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cluster"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicListingSummary"
                  },
                  "title": "Response Platform Directory Public Challenges Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/public/x/{company_slug}/{programme_slug}": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Listing",
        "operationId": "listing_public_x__company_slug___programme_slug__get",
        "parameters": [
          {
            "name": "company_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Company Slug"
            }
          },
          {
            "name": "programme_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Programme Slug"
            }
          },
          {
            "name": "projet_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Projet Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicListing"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/public/x/{company_slug}/{programme_slug}/apply": {
      "post": {
        "tags": [
          "public"
        ],
        "summary": "Apply",
        "description": "FR-201 to FR-208.\n\nSubmission is allowed with either consent declined (FR-202): declining is a\nreal choice, not a soft block. No password here \u2014 applicants sign in later\nvia the offer / set-password path once they have a reason to. A signed-in\napplicant may reuse the CV already on their profile.",
        "operationId": "apply_public_x__company_slug___programme_slug__apply_post",
        "parameters": [
          {
            "name": "company_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Company Slug"
            }
          },
          {
            "name": "programme_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Programme Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_apply_public_x__company_slug___programme_slug__apply_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/public/x/{company_slug}/{programme_slug}/notify-me": {
      "post": {
        "tags": [
          "public"
        ],
        "summary": "Notify Me",
        "description": "FR-103 \u2014 when applications are closed the CTA becomes a capture.",
        "operationId": "notify_me_public_x__company_slug___programme_slug__notify_me_post",
        "parameters": [
          {
            "name": "company_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Company Slug"
            }
          },
          {
            "name": "programme_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Programme Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotifyRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Notify Me Public X  Company Slug   Programme Slug  Notify Me Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublicListing": {
        "properties": {
          "company": {
            "type": "string",
            "title": "Company"
          },
          "company_slug": {
            "type": "string",
            "title": "Company Slug"
          },
          "company_logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Logo Url"
          },
          "programme_slug": {
            "type": "string",
            "title": "Programme Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "problem_statement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem Statement"
          },
          "deliverable": {
            "type": "string",
            "title": "Deliverable"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "applications_close_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applications Close At"
          },
          "start_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start At"
          },
          "submit_deadline_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submit Deadline At"
          },
          "pitch_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pitch At"
          },
          "seats_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seats Total"
          },
          "seats_remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seats Remaining"
          },
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/PublicCriterion"
            },
            "type": "array",
            "title": "Criteria"
          },
          "data_pack_preview": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Data Pack Preview"
          },
          "writeup_prompt": {
            "type": "string",
            "title": "Writeup Prompt"
          },
          "already_applied": {
            "type": "boolean",
            "title": "Already Applied",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "company",
          "company_slug",
          "company_logo_url",
          "programme_slug",
          "title",
          "role",
          "problem_statement",
          "deliverable",
          "state",
          "applications_close_at",
          "start_at",
          "submit_deadline_at",
          "pitch_at",
          "seats_total",
          "seats_remaining",
          "criteria",
          "data_pack_preview",
          "writeup_prompt"
        ],
        "title": "PublicListing",
        "description": "FR-101 \u2014 company, role, summary, dates, commitment, and what they get.\n\nThe full rubric renders here (FR-078): publishing it tells participants what\nthey are judged on and removes the \"it felt arbitrary\" complaint. There is\nno advantage in concealing it \u2014 the rubric rewards things that cannot be faked."
      },
      "PublicListingSummary": {
        "properties": {
          "company": {
            "type": "string",
            "title": "Company"
          },
          "company_slug": {
            "type": "string",
            "title": "Company Slug"
          },
          "company_logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Logo Url"
          },
          "programme_slug": {
            "type": "string",
            "title": "Programme Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "cluster": {
            "type": "string",
            "title": "Cluster"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "applications_close_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applications Close At"
          },
          "start_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start At"
          },
          "seats_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seats Total"
          },
          "seats_remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seats Remaining"
          }
        },
        "type": "object",
        "required": [
          "company",
          "company_slug",
          "company_logo_url",
          "programme_slug",
          "title",
          "role",
          "cluster",
          "state",
          "applications_close_at",
          "start_at",
          "seats_total",
          "seats_remaining"
        ],
        "title": "PublicListingSummary",
        "description": "FR-104/FR-105 \u2014 the row shape for both listing endpoints below.\n\nDeliberately thinner than `PublicListing`: a browsing stranger picking\nbetween programmes does not need the rubric or the data pack yet, only\nenough to decide whether to open the full page."
      },
...
    }
  }
}